requirements.txt
git+https://github.com/barretobrock/kavalkilu.git#egg=kavalkilu
git+https://github.com/barretobrock/pukr.git@main#egg=pukr
pandas
PyYaml
